Add functionality for different user roles and the access that the user gets depending on that role.
Remove the tabs interface from the index page and place it in a newly created profile page. Update the select statement in login-action.php to retrieve the role_id from the database. Also, add that role_id to a variable and use that in conditionals later. Use a nested if/statement to display tabs and tab navigation depending on the level of access each user is allowed.